requirements.txt
customtkinter==5.2.2
pytubefix==8.4.1
Pillow==11.0.0
requests==2.32.3